Change source for internal energy
The internal energy is no longer reported as part of the start up
of a phonon calculation (I think because the basis set correction
now gets stored in a .check file so does not need recalculating).
So need to use something else to get the internal energy and
calculate the helholtz free energy. However, the internal energy
reported elsewhere does not include the basis set correction. So
instead use the enthalpy and subtract PV remembering to do the
unit conversion to eV (from GPa and A^3).